TStr:NextToken: Rope in CharTypes into nexttok, as needed.
With this a basic flow wrt nexttok based on the new nexttoken based
mechanism should be done.

The flow needs to be thought throught and cross checked once still.

The drop_adjust logic and related potentially updated to avoid the
possible corner case, mentioned before.

This also needs to be thought through and cross checked once still.