TStr:NextToken: Add logic to handle escape sequences expand flag
The logic wrt accessing escseq map or not, if escseq mode is set,
is technically not required, as escseq mode will not be set if
escseq expand flag is not set. However just in case, in future,
if I change this relationship, so keeping the logic decoupled in
a brute force way for now.