TStr:NextTok+: Return partial token till then, if err
NextTok - returns the err string and tok till then as a tuple.

Other helpers built on top of NextTok, returns the err string,
which internally also contains the partial tok till then at the
end of the err msg.

The Err message contains tag at the beginning wrt the custom type
(TStr) and the helper method one is calling.