TStr: Esc the esc seq tag/marker ie slash, Testlib delim to ';'
Allow the escape sequence slash to be escaped and thus put in a
string so that it remains as such even after tokenisation.

Also update the testlib to use ';' as the delim, so that the full
string is treated as a single token, so that one can map the string
char position printed to be easily mappable when debugging.

Replace a system default escape sequence wrt \t to output newline.
To validate the possibility of replacing existing entries.

Also put the expected end string after calling tokenisation, in
the source, for reference, for future, when things may be all blurry.

Print the debug version of the generated token as well as the string
version, so that we can see the token with automatic escape sequencing
of special chars, as well as the effects of those special chars. This
should ehlp identify the effects of escape sequence processing, more
clearly.

Misc: Ustaad Hotel, Hridayam, Maheshinte Prathikaram